consolidate the type aliases in type.ts .
