+++ src/components/ui/Header.tsx
@@ -1,9 +1,6 @@
 'use client'
 
-type HeaderProps = {
-  labelName: React.ReactNode
-  className: string
-}
+import { HeaderProps } from '@/app/types/type'
 
 export const Header: React.FC<HeaderProps> = ({ labelName, className }) => {
   return <h1 className={className}>{labelName}</h1>
